Сделал модульную систему. Удалил файл functions.js
