+++ js/const.js
@@ -1,0 +1,50 @@
+const POST_COUNT = 25;
+const COMMENTS_COUNT = 2;
+const LIKES_MIN_COUNT = 15;
+const LIKES_MAX_COUNT = 200;
+const ID_MIN_COUNT = 1;
+const ID_MAX_COUNT = 25;
+const ID_COMMENT_MIN_COUNT = 1;
+const ID_COMMENT_MAX_COUNT = 99999;
+const NAMES = [
+  'Дарья',
+  'Вадим',
+  'Екатерина',
+  'Дмитрий',
+  'Елена',
+  'Сергей',
+  'Софья',
+  'Константин',
+  'Зоя',
+  'Александр'
+];
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const DESCRIPTIONS = [
+  'Эту фотографию я сделал сам! #me',
+  'С друзьями на шашлыках #отдых',
+  'Это мы на отдыхе #отдыхаемхорошо #веселоидружно',
+  'Собрались со всей семьей дома #милыйдом #отдыхаем #уютныйвечер',
+  'Еду домой! #милыйдом #уютныйвечер',
+  'Как здорово получилось! #Супер'
+];
+
+export {
+  POST_COUNT,
+  COMMENTS_COUNT,
+  LIKES_MIN_COUNT,
+  LIKES_MAX_COUNT,
+  ID_MIN_COUNT,
+  ID_MAX_COUNT,
+  ID_COMMENT_MIN_COUNT,
+  ID_COMMENT_MAX_COUNT,
+  NAMES,
+  COMMENTS,
+  DESCRIPTIONS
+};
